pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "waffle"
version = "0.0.1"
authors = [{ name = "chunan", email = "chunan.liu@outlook.com" }]
description = "A package for affinity ranking"
requires-python = ">=3.10"
dependencies = [
    "biopython==1.85",
    "numpy<2.0.0",
    "loguru>=0.7.2",
    "lightning==2.4.0",
    "lightning-flash==0.8.2",
    "lightning-utilities==0.11.9",
    "pytorch-lightning==1.8.6",
    "rootutils==1.0.7",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
]

[project.optional-dependencies]
# usage:
# pip install -e ".[dev]"  # install dev dependencies only
# pip install -e ".[dev,test,docs]"  # install dev, test, docs dependencies at once
dev = [
    "jupyter",
]
# usage: pip install -e ".[test]"
test = [
    "pytest",
]
# usage: pip install -e ".[docs]"
docs = [
    "sphinx",
    "mkdocs"
]

[tool.setuptools.packages.find]
include = ["waffle", "waffle.*"]
exclude = ["tests*"]
